second commit

+++ script.js
@@ -1,7 +1,7 @@
 function total()
 {
-        var num1 = parseInt(document.getElementById("hours").value);
-        var num2 = parseInt(document.getElementById("unit-price").value);
+        var num1 = parseFloat(document.getElementById("hours").value);
+        var num2 = parseFloat(document.getElementById("unit-price").value);
         document.getElementById("total").innerHTML = num1 * num2;
         document.getElementById("final").innerHTML = num1 * num2;
